update BlockDevice trait

+++ src/lib.rs
@@ -2,8 +2,8 @@
 //! ```rust
 //! pub trait BlockDevice {
 //!     type Error;
-//!     fn read(&self, buf: &mut [u8], address: usize, number_of_blocks: u32) -> Result<(), Self::Error>;
-//!     fn write(&self, buf: &[u8], address: usize, number_of_blocks: u32) -> Result<(), Self::Error>;
+//!     fn read(&self, buf: &mut [u8], address: usize, number_of_blocks: usize) -> Result<(), Self::Error>;
+//!     fn write(&self, buf: &[u8], address: usize, number_of_blocks: usize) -> Result<(), Self::Error>;
 //! }
 //! ```
 
@@ -12,6 +12,6 @@
 /// BlockDevice trait
 pub trait BlockDevice {
     type Error;
-    fn read(&self, buf: &mut [u8], address: usize, number_of_blocks: u32) -> Result<(), Self::Error>;
-    fn write(&self, buf: &[u8], address: usize, number_of_blocks: u32) -> Result<(), Self::Error>;
+    fn read(&self, buf: &mut [u8], address: usize, number_of_blocks: usize) -> Result<(), Self::Error>;
+    fn write(&self, buf: &[u8], address: usize, number_of_blocks: usize) -> Result<(), Self::Error>;
 }
